rename SvgTrail folder
